webhook: Support disabling sidecar for token refresh
The webhook adds a sidecar to the pod to refresh the generated token.
Running jobs with the sidecar never go to completion because the sidecar
keeps running when the main container of the job exits. See
https://github.com/doitintl/gtoken/issues/2.

Workaround this by adding an annotation to disable refreshing the token,
and as a result skip adding the sidecar.

Signed-off-by: Alex Tsitsimpis <alextsi@arrikto.com>